Implement reweighted MLE (Andrew & Kasy)
The newly added functions implement a maximum likelihood estimator for
the corrected population mean based on a likelihood function that was
reweighted by the publication probability of a specific result. The
publication probability is assumed to be a step function where
significant results have publication probability 1 and non-significant
results have publication probability p. If p is not set, a grid-search
over all possible ps and the provided theta values is performed (theta =
mu/sgm).